[Gold II] Title: 꿈틀꿈틀 호석 애벌레 - 효율성, Time: 16 ms, Memory: 3852 KB, Score: 47 point -BaekjoonHub

+++ 백준/Gold/20181. 꿈틀꿈틀 호석 애벌레 － 효율성/꿈틀꿈틀 호석 애벌레 － 효율성.cc
@@ -25,8 +25,8 @@ int main() {
         }
     }
 
-    int maxEnergy=select[0][0];
-    int dp[N]; dp[0]=select[0][0];
+    long long maxEnergy=select[0][0];
+    long long dp[N]; dp[0]=select[0][0];
     for(int i=1;i<N;i++) {
         dp[i]=dp[i-1];
         if(i-select[i][1]>=0) dp[i] = max(dp[i], dp[i-select[i][1]]+select[i][0]);
